move routering path from main to route
